Initial Draft
This script assumes that your device name is /dev/vda and you are running UEFI. In addition to this it assumes that your country of residence is India and your Locale preferences are set accordingly. Script is licensed under GLP V2 or later so totally free. Any feedback is welcome.
I have only tested successfully the UEFI install in a Virtual Box. BIOS Install still to be tested.